start a metabase conversion tool
This will be the tool that converts from the existing mysql-based
metabase instance to a newer, easier-to-manage raw test report. This may
need to be written a few times while we test a few databases out, but
the end goal of this is to have a better place to store test reports
that's easier to load so we can fix the most expensive page on the
entire cpantesters website.