start a module to migrate to a simpler metabase
The current cache of the metabase is compressed and encoded, and
unsuitable for anything but direct display of individual reports. We
want to eventually have a better metabase that enables searching and
querying over sets of reports.

This is mostly just a layout of the Beam::Runner container.

This is still a work in progress