fix metabase cache reports not being readable
If we allow MySQL to do its UTF-8 collation to Latin-1, it mangles the
binary data we put in the metabase cache (compressed Sereal and
compressed JSON). Then when we try to read it back out,
Data::FlexSerializer fails to understand and prints out an error.
However, that error looks fine, as my terminal is a UTF-8 terminal...

So we have to disable UTF-8 for the Metabase cache (and, indeed, all of
our connections. Not sure why we had it in the first place...)