add reprocessing of test reports
It's almost impossible to delete rows from the processed reports table:
InnoDB will go nuts trying to achieve locks on the rows, so the MySQL
docs suggest copying only the data you want to another table,
renaming the new table to the old table's name, and deleting the old
table. Tedious.

Instead, if we just have a flag to reprocess the report, we can update
the processed data in-place, saving us all time and effort.

This required a change to the schema module as well, so I updated the
prereq for that.