Use MySQL version 5.7 in Travis tests
... since this version supports the JSON field and hence would allow the
tests to pass.  Also, reinstalling MySQL solves a strange error which
looks like it's only associated with Travis and the MySQL version that
comes preinstalled on a build VM: when Test::mysqld tries to initialise,
it uses `mysql_install_db` internally, which for some reason tries to
find `my-default.cnf`, which doesn't exist, hence an error is thrown and
the tests fail.  If one installs one's own MySQL version, then this
issue disappears.